@@ -1,6 +1,11 @@
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import {
+  UntypedFormBuilder,
+  FormControl,
+  UntypedFormGroup,
+  Validators,
+} from '@angular/forms';
 import { Router } from '@angular/router';
 import { map, Observable, shareReplay } from 'rxjs';
 import { AuthService, User } from '../auth.service';
@@ -13,38 +18,44 @@ import { AuthService, User } from '../auth.service';
 export class LoginComponent implements OnInit {
   hide = true;
 
-  loginForm: FormGroup;
+  loginForm: UntypedFormGroup;
 
-  isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
-  .pipe(
-    map(result => result.matches),
-    shareReplay()
-  );
+  isHandset$: Observable<boolean> = this.breakpointObserver
+    .observe(Breakpoints.Handset)
+    .pipe(
+      map((result) => result.matches),
+      shareReplay()
+    );
 
   constructor(
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private auth: AuthService,
     private breakpointObserver: BreakpointObserver,
-    private router: Router,
+    private router: Router
   ) {
     this.loginForm = this.fb.group({
-      email: [null, [Validators.required, Validators.pattern('[a-z0-9]+@[a-z]+\.[a-z]{2,3}')]],
+      email: [
+        null,
+        [
+          Validators.required,
+          Validators.pattern('[a-z0-9]+@[a-z]+.[a-z]{2,3}'),
+        ],
+      ],
       password: [null, [Validators.required, Validators.minLength(8)]],
     });
-   }
-
-  ngOnInit(): void {
-
   }
+
+  ngOnInit(): void {}
 
   onSubmit(): void {
     const { email, password } = this.loginForm.value;
-    this.auth.login({email, password}).subscribe({next: ()=>{
-      this.router.navigateByUrl('/pages/login');
-    },
-    error: ()=> {
-      console.log('Retry');
-    }});
+    this.auth.login({ email, password }).subscribe({
+      next: () => {
+        this.router.navigateByUrl('/pages/login');
+      },
+      error: () => {
+        console.log('Retry');
+      },
+    });
   }
-
 }
