@@ -1,21 +1,43 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AuthComponent } from './auth.component';
+import { ForgotComponent } from './forgot/forgot.component';
 import { LoginComponent } from './login/login.component';
+import { LogoutComponent } from './logout/logout.component';
+import { RegisterComponent } from './register/register.component';
+import { ResetComponent } from './reset/reset.component';
 
-const routes: Routes = [{
-  path: '',
-  component: AuthComponent,
-  children: [
-    {
-      path: 'login',
-      component: LoginComponent,
-    },
-  ]
-}];
+const routes: Routes = [
+  {
+    path: '',
+    component: AuthComponent,
+    children: [
+      {
+        path: 'login',
+        component: LoginComponent,
+      },
+      {
+        path: 'register',
+        component: RegisterComponent,
+      },
+      {
+        path: 'forgot',
+        component: ForgotComponent,
+      },
+      {
+        path: 'reset',
+        component: ResetComponent,
+      },
+      {
+        path: 'logout',
+        component: LogoutComponent,
+      },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class AuthRoutingModule { }
+export class AuthRoutingModule {}
