@@ -1,15 +1,49 @@
+import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
+import {
+  UntypedFormBuilder,
+  FormControl,
+  UntypedFormGroup,
+  Validators,
+} from '@angular/forms';
+import { Router } from '@angular/router';
+import { map, Observable, shareReplay } from 'rxjs';
+import { AuthService, User } from '../auth.service';
 
 @Component({
   selector: 'app-register',
   templateUrl: './register.component.html',
-  styleUrls: ['./register.component.scss']
+  styleUrls: ['./register.component.scss'],
 })
 export class RegisterComponent implements OnInit {
+  hide = true;
 
-  constructor() { }
+  registerForm: UntypedFormGroup;
 
-  ngOnInit(): void {
+  constructor(
+    private fb: UntypedFormBuilder,
+    private auth: AuthService,
+    private router: Router
+  ) {
+    this.registerForm = this.fb.group({
+      fullName: [null, [Validators.required, Validators.minLength(6)]],
+      company: [null],
+      email: [
+        null,
+        [
+          Validators.required,
+          Validators.pattern('[a-z0-9]+@[a-z]+.[a-z]{2,3}'),
+        ],
+      ],
+      password: [null, [Validators.required, Validators.minLength(8)]],
+      repassword: [null, [Validators.required, Validators.minLength(8)]],
+    });
   }
 
+  ngOnInit(): void {}
+
+  onSubmit(): void {
+    const user = this.registerForm.value;
+    this.auth.register(user).subscribe();
+  }
 }
