@@ -4,7 +4,7 @@ import {
   HttpHandler,
   HttpEvent,
   HttpInterceptor,
-  HTTP_INTERCEPTORS
+  HTTP_INTERCEPTORS,
 } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { TokenDomains, TokenService } from './token.service';
@@ -13,16 +13,18 @@ export const TOKEN_HEADER = 'Authorization';
 
 @Injectable()
 export class TokenInterceptor implements HttpInterceptor {
+  constructor(private token: TokenService) {}
 
-  constructor(
-    private token: TokenService,
-  ) {}
-
-  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
+  intercept(
+    request: HttpRequest<unknown>,
+    next: HttpHandler
+  ): Observable<HttpEvent<unknown>> {
     let authRequest = request;
     const token = this.token.retrieveToken(TokenDomains.OrbAPI);
-    if(token !==null) {
-      authRequest = request.clone({headers: request.headers.set(TOKEN_HEADER, 'Bearer '+token)});
+    if (token !== null) {
+      authRequest = request.clone({
+        headers: request.headers.set(TOKEN_HEADER, 'Bearer ' + token),
+      });
     }
 
     return next.handle(authRequest);
@@ -30,5 +32,5 @@ export class TokenInterceptor implements HttpInterceptor {
 }
 
 export const authInterceptorProviders = [
-  { provide: HTTP_INTERCEPTORS, useClass: TokenInterceptor, multi: true }
+  { provide: HTTP_INTERCEPTORS, useClass: TokenInterceptor, multi: true },
 ];
