@@ -2,29 +2,27 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Event, NavigationEnd, Router } from '@angular/router';
 import { distinctUntilChanged, filter } from 'rxjs';
 
-
 @Component({
   selector: 'app-breadcrumbs',
   templateUrl: './breadcrumbs.component.html',
-  styleUrls: ['./breadcrumbs.component.scss']
+  styleUrls: ['./breadcrumbs.component.scss'],
 })
 export class BreadcrumbsComponent implements OnInit {
-  public breadcrumbs: IBreadCrumb[]
+  public breadcrumbs: IBreadCrumb[];
 
-  constructor(
-    private router: Router,
-    private activatedRoute: ActivatedRoute,
-  ) {
+  constructor(private router: Router, private activatedRoute: ActivatedRoute) {
     this.breadcrumbs = this.buildBreadCrumb(this.activatedRoute.root);
   }
 
   ngOnInit() {
-    this.router.events.pipe(
-      filter((event: Event) => event instanceof NavigationEnd),
-      distinctUntilChanged(),
-  ).subscribe(() => {
-      this.breadcrumbs = this.buildBreadCrumb(this.activatedRoute.root);
-  })
+    this.router.events
+      .pipe(
+        filter((event: Event) => event instanceof NavigationEnd),
+        distinctUntilChanged()
+      )
+      .subscribe(() => {
+        this.breadcrumbs = this.buildBreadCrumb(this.activatedRoute.root);
+      });
   }
 
   /**
@@ -33,48 +31,55 @@ export class BreadcrumbsComponent implements OnInit {
    * @param url
    * @param breadcrumbs
    */
-/**
- * Recursively build breadcrumb according to activated route.
- * @param route
- * @param url
- * @param breadcrumbs
- */
- buildBreadCrumb(route: ActivatedRoute, url: string = '', breadcrumbs: IBreadCrumb[] = []): IBreadCrumb[] {
-  //If no routeConfig is avalailable we are on the root path
-  let { routeConfig } = route;
-  let { data, path } = { path: '', data: { breadcrumb: '', disabled: false}, ...routeConfig};
-  let { breadcrumb: label, disabled } = data;
+  /**
+   * Recursively build breadcrumb according to activated route.
+   * @param route
+   * @param url
+   * @param breadcrumbs
+   */
+  buildBreadCrumb(
+    route: ActivatedRoute,
+    url: string = '',
+    breadcrumbs: IBreadCrumb[] = []
+  ): IBreadCrumb[] {
+    //If no routeConfig is avalailable we are on the root path
+    let { routeConfig } = route;
+    let { data, path } = {
+      path: '',
+      data: { breadcrumb: '', disabled: false },
+      ...routeConfig,
+    };
+    let { breadcrumb: label, disabled } = data;
 
+    // If the route is dynamic route such as ':id', remove it
+    const lastRoutePart = path.split('/').pop() || '';
+    const isDynamicRoute = lastRoutePart.startsWith(':');
+    if (isDynamicRoute && !!route.snapshot) {
+      const paramName = lastRoutePart.split(':')[1];
+      path = path.replace(lastRoutePart, route.snapshot.params[paramName]);
+      label = route.snapshot.params[paramName];
+    }
 
+    //In the routeConfig the complete path is not available,
+    //so we rebuild it each time
+    const nextUrl = path ? `${url}/${path}` : url;
 
-  // If the route is dynamic route such as ':id', remove it
-  const lastRoutePart = path.split('/').pop() || '';
-  const isDynamicRoute = lastRoutePart.startsWith(':');
-  if(isDynamicRoute && !!route.snapshot) {
-    const paramName = lastRoutePart.split(':')[1];
-    path = path.replace(lastRoutePart, route.snapshot.params[paramName]);
-    label = route.snapshot.params[paramName];
-  }
-
-  //In the routeConfig the complete path is not available,
-  //so we rebuild it each time
-  const nextUrl = path ? `${url}/${path}` : url;
-
-  const breadcrumb: IBreadCrumb = {
+    const breadcrumb: IBreadCrumb = {
       label: label,
       url: nextUrl,
       disabled,
-  };
-  // Only adding route with non-empty label
-  const newBreadcrumbs = breadcrumb.label ? [ ...breadcrumbs, breadcrumb ] : [ ...breadcrumbs];
-  if (route.firstChild) {
+    };
+    // Only adding route with non-empty label
+    const newBreadcrumbs = breadcrumb.label
+      ? [...breadcrumbs, breadcrumb]
+      : [...breadcrumbs];
+    if (route.firstChild) {
       //If we are not on our current path yet,
       //there will be more children to look after, to build our breadcumb
       return this.buildBreadCrumb(route.firstChild, nextUrl, newBreadcrumbs);
+    }
+    return newBreadcrumbs;
   }
-  return newBreadcrumbs;
-}
-
 }
 
 export interface IBreadCrumb {
