@@ -4,15 +4,18 @@ import { Sink } from 'src/app/services/interfaces/sink.interface';
 @Component({
   selector: 'app-sink-item',
   templateUrl: './sink-item.component.html',
-  styleUrls: ['./sink-item.component.scss']
+  styleUrls: ['./sink-item.component.scss'],
 })
 export class SinkItemComponent implements OnInit {
   @Input()
   sink: Sink = {};
 
-  constructor() { }
+  constructor() {}
 
-  ngOnInit(): void {
+  ngOnInit(): void {}
+
+  sinkState() {
+    const state = 'stale';
+    return state;
   }
-
 }
