@@ -6,18 +6,14 @@ import { OrbService } from 'src/app/services/orb.service';
 @Component({
   selector: 'app-datasets',
   templateUrl: './datasets.component.html',
-  styleUrls: ['./datasets.component.scss']
+  styleUrls: ['./datasets.component.scss'],
 })
 export class DatasetsComponent implements OnInit {
   datasets$: Observable<Dataset[]>;
 
-  constructor(
-    private orb: OrbService,
-  ) {
+  constructor(private orb: OrbService) {
     this.datasets$ = this.orb.getDatasetListView();
   }
 
-  ngOnInit(): void {
-  }
-
+  ngOnInit(): void {}
 }
