@@ -2,10 +2,9 @@ import { KeyValue } from '@angular/common';
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-  name: 'tagColor'
+  name: 'tagColor',
 })
 export class TagColorPipe implements PipeTransform {
-
   /**
    * <`hsl(${h}deg, 90%, 65%)`>
    * @param tagList {<KeyValue<string,string>> | string}
@@ -17,12 +16,12 @@ export class TagColorPipe implements PipeTransform {
       const h = Math.abs(
         `${value}}`
           .split('')
-          .map(v => v.charCodeAt(0))
-          .reduce((a, v) => a + ((a << 7) + (a << 3)) ^ v) % 360);
+          .map((v) => v.charCodeAt(0))
+          .reduce((a, v) => (a + ((a << 7) + (a << 3))) ^ v) % 360
+      );
       return `hsl(${h}, 90%, 65%)`;
     }
 
     return 'transparent';
   }
-
 }
