@@ -6,13 +6,10 @@ import { Dataset } from './interfaces/dataset.interface';
 import { OrbPagination } from './interfaces/pagination.interface';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class DatasetService {
-
-  constructor(
-    private http: HttpClient,
-  ) { }
+  constructor(private http: HttpClient) {}
 
   addDataset(dataset: Dataset) {
     return this.http.post(environment.datasets, dataset);
@@ -28,26 +25,30 @@ export class DatasetService {
   }
 
   getAllDatasets() {
-    let page = {order: 'name', dir: 'asc', limit: 100, offset: 0} as OrbPagination<Dataset>;
+    let page = {
+      order: 'name',
+      dir: 'asc',
+      limit: 100,
+      offset: 0,
+    } as OrbPagination<Dataset>;
 
-    return this.getDatasetPage(page)
-      .pipe(
-        expand(page => {
-          return page.next ? this.getDatasetPage(page.next) : EMPTY;
-        }),
-        delay(100),
-        reduce((acc, value) => {
-          acc.data = [...acc?.data || [], ...value?.data || []];
-          acc.offset = 0;
-          acc.total = acc.data.length;
-          return acc;
-        }, page),
-        map(page => page.data),
-      );
+    return this.getDatasetPage(page).pipe(
+      expand((page) => {
+        return page.next ? this.getDatasetPage(page.next) : EMPTY;
+      }),
+      delay(100),
+      reduce((acc, value) => {
+        acc.data = [...(acc?.data || []), ...(value?.data || [])];
+        acc.offset = 0;
+        acc.total = acc.data.length;
+        return acc;
+      }, page),
+      map((page) => page.data)
+    );
   }
 
   getDatasetPage(page: OrbPagination<Dataset>) {
-    const {order, dir, offset, limit} = page;
+    const { order, dir, offset, limit } = page;
 
     let params = new HttpParams()
       .set('order', order)
@@ -55,15 +56,26 @@ export class DatasetService {
       .set('offset', offset.toString())
       .set('limit', limit.toString());
 
-    return this.http.get(`${environment.datasets}`, {params})
-    .pipe(map((resp: any) => {
-      const {order, dir, offset, limit, total, datasets} = resp;
-      const next = offset + limit < total && {
-        limit, order, dir,
-        offset: (parseInt(offset, 10) + parseInt(limit, 10)).toString(),
-      }
-      return {order, dir, offset, limit, total, data: datasets, next} as OrbPagination<Dataset>;
-    }));
+    return this.http.get(`${environment.datasets}`, { params }).pipe(
+      map((resp: any) => {
+        const { order, dir, offset, limit, total, datasets } = resp;
+        const next = offset + limit < total && {
+          limit,
+          order,
+          dir,
+          offset: (parseInt(offset, 10) + parseInt(limit, 10)).toString(),
+        };
+        return {
+          order,
+          dir,
+          offset,
+          limit,
+          total,
+          data: datasets,
+          next,
+        } as OrbPagination<Dataset>;
+      })
+    );
   }
 
   deleteDataset(id: string) {
